Logs and Exceptions.

+++ BadCodeTestApp/Commands/FileCommands/Create.cs
@@ -1,3 +1,4 @@
+using System;
 using System.IO;
 
 namespace BadCodeTestApp.Commands.FileCommands
@@ -11,6 +12,14 @@ namespace BadCodeTestApp.Commands.FileCommands
 
         protected override void ExCommand(string[] args)
         {
+            if (File.Exists(args[1]))
+            {
+                Program.Logger.Debug("File " + args[1] + " is already exist.");
+                Console.WriteLine("The file is already exist.");
+                return;
+            }
+
+            Program.Logger.Debug("File " + args[1] + " create.");
             File.Create(args[1]);
         }
     }
